Combine fluctuation tests
