+++ test/CMakeLists.txt
@@ -228,7 +228,6 @@ celeritas_add_test(celeritas/Constants.test.cc ${_optional_geant4_link})
 set(CELERITASTEST_PREFIX celeritas/em)
 celeritas_add_test(celeritas/em/BetheHeitler.test.cc)
 celeritas_add_test(celeritas/em/CombinedBrem.test.cc)
-celeritas_add_test(celeritas/em/EnergyLossFluctuation.test.cc)
 celeritas_add_test(celeritas/em/EPlusGG.test.cc)
 celeritas_add_test(celeritas/em/Fluctuation.test.cc)
 celeritas_add_test(celeritas/em/KleinNishina.test.cc)
